- Persistence Leaderboard v2.0
Changed variables of player data class to lists in order to save multiple players and implemented complete mechanism from adding new player data to updating current player data and removing player from the list if no space in the list and if the current player's score is higher then the lowest scorer in the list. Actually leaderboard is still not implemented yet.